+++ webpack-no-rdflib.config.js
@@ -1,0 +1,8 @@
+var path = require('path')
+
+var config = require('./webpack.config')
+
+// Exclude rdflib
+config.externals['rdflib'] = '$rdf'
+
+module.exports = config